refactor - split header file into dicomm and jwm
