+++ src/messages/headers/mod.rs
@@ -1,9 +1,11 @@
-mod headers;
+mod didcomm;
 mod jwk;
+mod jwm;
 mod prior_claims;
 mod types;
 
-pub use headers::*;
+pub use didcomm::*;
 pub use jwk::*;
+pub use jwm::*;
 pub use prior_claims::*;
 pub use types::*;
